pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "shotsieve"
version = "0.2.0"
description = "Local-first photo quality scanner and cache"
readme = "README.md"
license = {file = "LICENSE"}
requires-python = ">=3.11"
classifiers = [
	"License :: OSI Approved :: GNU Affero General Public License v3 or later (AGPLv3+)",
	"Programming Language :: Python :: 3",
	"Programming Language :: Python :: 3 :: Only",
]
dependencies = [
	"numpy>=2.0.0",
	"Pillow>=11.0.0",
]

[project.optional-dependencies]
test = [
	"pytest>=8.0",
	"playwright>=1.58.0",
	"ruff>=0.15.0",
]
lint = [
	"ruff>=0.15.0",
]
format-loaders = [
	"pillow-heif>=1.3.0",
	"rawpy>=0.26.0",
]
learned-iqa = [
	"pyiqa>=0.1.15",
	"icecream>=2.1.3",
]
learned-iqa-directml = [
	"pyiqa>=0.1.15",
	"icecream>=2.1.3",
	"torch-directml>=0.2.5.dev0; platform_system == 'Windows' and python_version < '3.13'",
]
windows-build = [
	"setuptools<81",
	"pyinstaller>=6.19",
]

[project.scripts]
shotsieve-desktop = "shotsieve.desktop:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
shotsieve = ["static/*.html", "static/*.js", "static/*.css"]